requirements.txt
streamlit
PyPDF2
python-docx
docx2txt
tools
ollama